Changes to expose simple message creation in corresponding JSON service
