+++ kura/distrib/config/kura.build.properties
@@ -8,7 +8,7 @@ org.eclipse.kura.broker.artemis.simple.mqtt.version=1.0.200
 org.eclipse.kura.broker.artemis.xml.version=1.0.200
 org.eclipse.kura.core.version=1.0.500-SNAPSHOT
 org.eclipse.kura.core.certificates.version=1.0.400
-org.eclipse.kura.core.cloud.version=1.1.300
+org.eclipse.kura.core.cloud.version=1.1.400-SNAPSHOT
 org.eclipse.kura.core.comm.version=1.0.400
 org.eclipse.kura.core.configuration.version=2.0.200-SNAPSHOT
 org.eclipse.kura.core.crypto.version=1.0.400
@@ -90,11 +90,11 @@ org.eclipse.kura.rest.provider.version=1.0.200
 org.eclipse.kura.rest.asset.provider.version=1.0.200
 org.eclipse.kura.hook.file.move.provider.version=1.0.200
 org.eclipse.kura.misc.cloudcat.version=1.0.200
-org.eclipse.kura.json.marshaller.unmarshaller.provider.version=1.0.100
+org.eclipse.kura.json.marshaller.unmarshaller.provider.version=1.0.200-SNAPSHOT
 org.eclipse.kura.xml.marshaller.unmarshaller.provider.version=1.0.100
 org.eclipse.kura.wire.component.conditional.provider.version=1.0.100
 org.eclipse.kura.wire.component.join.provider.version=1.0.100
-org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider.version=1.0.0
+org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider.version=1.0.100-SNAPSHOT
 
 ## features versions
 org.eclipse.kura.driver.opcua.version=1.0.400-SNAPSHOT
