+++ kura/org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider/META-INF/MANIFEST.MF
@@ -26,6 +26,7 @@ Import-Package: com.eclipsesource.json;version="0.9.4",
  org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,1.1)",
+ org.eclipse.kura.marshalling;version="[1.0,2.0)",
  org.eclipse.kura.message;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[2.0,3.0)",
  org.eclipse.kura.net.modem;version="[2.0,3.0)",
