+++ kura/org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider/src/main/java/org/eclipse/kura/internal/cloudconnection/eclipseiot/mqtt/cloud/CloudConnectionManagerImpl.java
@@ -19,6 +19,7 @@ import static org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.
 import static org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.MessageConstants.RETAIN;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.Dictionary;
 import java.util.HashMap;
@@ -55,6 +56,8 @@ import org.eclipse.kura.core.data.DataServiceImpl;
 import org.eclipse.kura.data.DataService;
 import org.eclipse.kura.data.listener.DataServiceListener;
 import org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.MessageType;
+import org.eclipse.kura.marshalling.Marshaller;
+import org.eclipse.kura.marshalling.Unmarshaller;
 import org.eclipse.kura.message.KuraApplicationTopic;
 import org.eclipse.kura.message.KuraPayload;
 import org.eclipse.kura.net.NetworkService;
@@ -98,6 +101,8 @@ public class CloudConnectionManagerImpl
     private PositionService positionService;
     private EventAdmin eventAdmin;
     private CertificatesService certificatesService;
+    private Unmarshaller jsonUnmarshaller;
+    private Marshaller jsonMarshaller;
 
     // package visibility for LifeCyclePayloadBuilder
     String imei;
@@ -197,6 +202,22 @@ public class CloudConnectionManagerImpl
 
     public void unsetEventAdmin(EventAdmin eventAdmin) {
         this.eventAdmin = null;
+    }
+    
+    public void setJsonUnmarshaller(Unmarshaller jsonUnmarshaller) {
+        this.jsonUnmarshaller = jsonUnmarshaller;
+    }
+
+    public void unsetJsonUnmarshaller(Unmarshaller jsonUnmarshaller) {
+        this.jsonUnmarshaller = null;
+    }
+
+    public void setJsonMarshaller(Marshaller jsonMarshaller) {
+        this.jsonMarshaller = jsonMarshaller;
+    }
+
+    public void unsetJsonMarshaller(Marshaller jsonMarshaller) {
+        this.jsonMarshaller = null;
     }
 
     // ----------------------------------------------------------------
@@ -409,13 +430,16 @@ public class CloudConnectionManagerImpl
         KuraPayload kuraPayload = null;
 
         if (this.options.getPayloadEncoding() == SIMPLE_JSON) {
-            kuraPayload = createKuraPayloadFromJson(payload);
+            try {
+                kuraPayload = createKuraPayloadFromJson(payload);
+            } catch (KuraException e) {
+                logger.warn("Error creating Kura Payload from Json", e);
+            }
         } else if (this.options.getPayloadEncoding() == KURA_PROTOBUF) {
             kuraPayload = createKuraPayloadFromProtoBuf(topic, payload);
         }
 
         try {
-
             boolean validMessage = isValidMessage(kuraTopic, kuraPayload);
 
             if (validMessage) {
@@ -639,12 +663,12 @@ public class CloudConnectionManagerImpl
         return bytes;
     }
 
-    private byte[] encodeJsonPayload(KuraPayload payload) {
-        return CloudPayloadJsonEncoder.getBytes(payload);
+    private byte[] encodeJsonPayload(KuraPayload payload) throws KuraException {
+        return this.jsonMarshaller.marshal(payload).getBytes(StandardCharsets.UTF_8);
     }
 
-    private KuraPayload createKuraPayloadFromJson(byte[] payload) {
-        return CloudPayloadJsonDecoder.buildFromByteArray(payload);
+    private KuraPayload createKuraPayloadFromJson(byte[] payload) throws KuraException {
+        return this.jsonUnmarshaller.unmarshal(new String(payload), KuraPayload.class);
     }
 
     private KuraPayload createKuraPayloadFromProtoBuf(String topic, byte[] payload) {
