+++ kura/org.eclipse.kura.core.cloud/META-INF/MANIFEST.MF
@@ -30,6 +30,7 @@ Import-Package: com.eclipsesource.json;version="0.9.4",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,1.1)",
+ org.eclipse.kura.marshalling;version="[1.0,2.0)",
  org.eclipse.kura.message;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[2.0,3.0)",
  org.eclipse.kura.net.modem;version="[2.0,3.0)",
