+++ kura/org.eclipse.kura.json.marshaller.unmarshaller.provider/META-INF/MANIFEST.MF
@@ -11,8 +11,10 @@ Import-Package: com.eclipsesource.json;version="0.9.4",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.core.configuration;version="[2.0,3.0)",
  org.eclipse.kura.marshalling;version="[1.0,1.1)",
+ org.eclipse.kura.message;version="1.2.0",
  org.eclipse.kura.wire;version="[2.0,3.0)",
- org.eclipse.kura.wire.graph;version="[1.0,2.0)"
+ org.eclipse.kura.wire.graph;version="[1.0,2.0)",
+ org.slf4j;version="1.7.25"
 Service-Component: OSGI-INF/*.xml
 Bundle-ActivationPolicy: lazy
 Export-Package: org.eclipse.kura.internal.json.marshaller.unmarshaller;version="1.0.0"
