+++ kura/test/org.eclipse.kura.core.test/pom.xml
@@ -313,6 +313,11 @@
                             <autoStart>true</autoStart>
                         </bundle>
                         <bundle>
+                            <id>org.eclipse.kura.json.marshaller.unmarshaller.provider</id>
+                            <level>4</level>
+                            <autoStart>true</autoStart>
+                        </bundle>
+                        <bundle>
                             <id>org.eclipse.kura.core.configuration</id>
                             <level>4</level>
                             <autoStart>true</autoStart>
@@ -547,6 +552,11 @@
                             </requirement>
                             <requirement>
                                 <type>eclipse-plugin</type>
+                                <id>org.eclipse.kura.json.marshaller.unmarshaller.provider</id>
+                                <versionRange>0.0.0</versionRange>
+                            </requirement>
+                            <requirement>
+                                <type>eclipse-plugin</type>
                                 <id>org.eclipse.kura.core.configuration</id>
                                 <versionRange>0.0.0</versionRange>
                             </requirement>
@@ -583,11 +593,6 @@
                             <requirement>
                                 <type>eclipse-plugin</type>
                                 <id>org.eclipse.kura.emulator</id>
-                                <versionRange>0.0.0</versionRange>
-                            </requirement>
-                            <requirement>
-                                <type>eclipse-plugin</type>
-                                <id>org.eclipse.kura.xml.marshaller.unmarshaller.provider</id>
                                 <versionRange>0.0.0</versionRange>
                             </requirement>
                             <requirement>
