+++ kura/test/org.eclipse.kura.json.marshaller.unmarshaller.provider.test/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.core.testutil;version="[1.0,2.0)",
  org.eclipse.kura.internal.json.marshaller.unmarshaller;version="[1.0,1.1)",
+ org.eclipse.kura.message;version="[1.2,2.0)",
  org.eclipse.kura.wire;version="[2.0,3.0)",
  org.eclipse.kura.wire.graph;version="[1.0,2.0)",
  org.junit;version="4.12.0"
